ref(team-roles) Move OrganizationTeamMemberDetailsEndpoint (#36549)
