+++ src/sentry/apidocs/public_exclusion_list.py
@@ -223,6 +223,9 @@ from sentry.api.endpoints.organization_member import (
     OrganizationMemberDetailsEndpoint,
     OrganizationMemberIndexEndpoint,
 )
+from sentry.api.endpoints.organization_member.team_details import (
+    OrganizationMemberTeamDetailsEndpoint,
+)
 from sentry.api.endpoints.organization_member_issues_assigned import (
     OrganizationMemberIssuesAssignedEndpoint,
 )
@@ -231,9 +234,6 @@ from sentry.api.endpoints.organization_member_issues_bookmarked import (
 )
 from sentry.api.endpoints.organization_member_issues_viewed import (
     OrganizationMemberIssuesViewedEndpoint,
-)
-from sentry.api.endpoints.organization_member_team_details import (
-    OrganizationMemberTeamDetailsEndpoint,
 )
 from sentry.api.endpoints.organization_member_unreleased_commits import (
     OrganizationMemberUnreleasedCommitsEndpoint,
